requirements.txt
discord.py
requests
python-dotenv
pytest
pytest-asyncio
logging
